+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 待发行
 
-1. 切换为 revolt/event-loop 组件实现事件调度
+1. 事件调度更改为 `revolt/event-loop` 组件
 2. 参考 Laravel 更改目录以及实现代码实现
+3. `DI` 更换为 `symfony/dependency-injection`
 
 ## v0.0.2